get_actions() is changed to sample() for action_space

+++ babopy/utils.py
@@ -6,10 +6,10 @@ algo_map = {
     'random': RandomAlgorithm
 }
 
-def get_random_state(action: int) -> typing.Tuple[float, float, bool, bool, dict]:
+def get_random_state(action: int) -> typing.Tuple[int, float, bool, bool, dict]:
     """
     Returns a random state.
     :return: A random state
-    :rtype: typing.Tuple[float, float, bool, bool, dict]
+    :rtype: typing.Tuple[int, float, bool, bool, dict]
     """
     return random.random(), random.random(), random.choice([True, False]), random.choice([True, False]), {"event": "A random state is sent."}
